fix: Fix -H argument, add other missing arguments
* Argument `-H` follows only symlinks supplied on the command-line.
* Add arguments `-a`, `-d`, `-L`, `-P`.
* If arguments end with a slash (/), the children of such arguments (directories) will be compied instead of the argument (directory) itself.

BREAKING CHANGE: The argument `-H` dereferences only symlinks supplied on the command-line. Not symlinks found in the subdirectories any more. If you want to dereference all symlinks, use the argument `L`. Also, if an argument ends with a slash (/), its children will be copied instead. If you want to copy the whole directory, do not end its name on the command line by a slash (/).